demo/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==3.2.10
djangorestframework==3.13.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
Werkzeug==2.2.2
